Update Main.java
Looped player's input to actually let them play the game. Hid the tiles content until it's revealed. Display the board on each input to see the changes. Made a displayBoard function which displays the board and playerInput function to ask for player input and make the tile's isClicked variable to true